+++ eve/tools/media_utils/subsession/test.json
@@ -1,0 +1,6 @@
+{   
+    "prompt": "A cat in outer space wearing sunglasses and a sombrero",
+    "init_image": "https://d14i3advvh2bvd.cloudfront.net/48f4b354bd5711b4d2234a9b8f05b193e186df8639ca72273c068e8b4f1910f1.png",
+    "n_samples": 2,
+    "aspect_ratio": "16:9"
+}
